+++ server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const app = express();
 const path = require('path');
 
-const PORT = process.env.PORT || 8080;
+const PORT = process.env.PORT || 80;
 
 app.use(express.static(path.join(__dirname, 'build')));
