+++ src/index.html
@@ -6,11 +6,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="stylesheet" href="vendor.css" />
     <link rel="stylesheet" href="app.css" />
+    <style>
+      @keyframes spin { 100% { transform: rotate(360deg); } }
+    </style>
   </head>
 
-  <body ng-app="app" ng-strict-di style="margin: 0; overflow: hidden;">
+  <body ng-app="app" ng-strict-di style="margin: 0; overflow: hidden; background: #000;">
     <div class="container">
-      <canvas id="canvas"></canvas>
+      <canvas id="canvas" style="animation:spin 10s linear infinite;"></canvas>
     </div>
 
     <script src="vendor.js"></script>
